[feat] 홈 페이지 및 컴포넌트 수정
